requirements.txt
matplotlib~=3.5.1
pandas~=1.4.1
python-rake~=1.5.0